+++ src/modules/admin/admin.tsx
@@ -1,0 +1,27 @@
+'use client';
+
+import { SC } from '@/modules/admin/admin.styles';
+import { useAuth } from '@/shared/context/auth';
+import useTranslation from 'next-translate/useTranslation';
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+
+const AdminPanel = () => {
+  const router = useRouter();
+  const { username } = useAuth();
+  const { t } = useTranslation('admin');
+
+  useEffect(() => {
+    if (!username) {
+      router.push('/login');
+    }
+  }, [username, router]);
+
+  return (
+    <SC.Container>
+      <SC.Title>Адмін панель</SC.Title>
+    </SC.Container>
+  );
+};
+
+export default AdminPanel;
