+++ tsconfig.json
@@ -28,8 +28,6 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "src/shared/context/auth.jsx",
-    "src/shared/context/local-storage.tsx"
   ],
   "exclude": ["node_modules"]
 }
